Add files via upload
I noted that there is a way that I can parse the json file into php by decoding it. I have tried to do it, with some searching and tutorials from google , however I have failed. So I decided to do it manually in html, while still following the requirements for the output.